Two changes are being added:
a. Empty frame command array for empty frames will not be generated any more.
In addition, a frameCount will be dumped in the JSON for a timeline. This denotes the total number of frames in the timeline.
This change is being done to reduce the JSON size.

b. We use a path separator during creation of runtime folder.

Right now, only the Win binary is being pushed along with source file changes.