Multiple changes are done part of this changelist:
1. Added an option "Compact Data" in the publish settings. This can be used by the user to configure the desired precision of the floating point numbers.
By default, this option is on and is set to "High" which means precision of 3 (maximum of 3 digits after decimal point) will be used.
This results in significant reduction of JSON size. If this option is turned off, then the precision of 6 is used.
Also, trailing zeroes in the floating point numbers are removed.

2. Changed the path data to use relative co-ordinates for Curveto (bezier curve) and LineTo commands.
Using relative co-ordinates generally reduce the JSON file size.

3. Unnecessary spaces in the path string has been removed if "Minify" option is turned on.