Fix for issue#55 - Added "Overwite HTML" option in the publish settings. This option is turned on by default.
Option is ON - HTML is always overwritten.
Option is OFF and output HTML exists - HTML is not overwritten.
Option is OFF and output HTML does not exist - HTML is created.

Apart from the above change, a new "Save" button is added in the publish settings.
This can be used to save the publish settings (w/o publish).
Also, made the file output path optional. User can leave it blank.

MAC binary and zxp will be added in separately.